rename test_datbase to test_db_manager, tests for crud operations, add relation for role to collab, add rich lib

+++ src/core/db_service.py
@@ -1,51 +1,52 @@
 from sqlalchemy import select, inspect
 
-from core.db_manager import DBManager
-
 
 class DBService:
-    def __init__(self):
-        self.session = DBManager().get_session()
+    def __init__(self, session):
+        self.session = session
 
-    def create(self, model_class, data: dict) -> None:
+    def create(self, model_class, data: dict) -> bool:
         obj = model_class(**data)
 
         with self.session.begin() as session:
             session.add(obj)
+            return True
+        return False
 
-    def get_all(self, model_class) -> list[dict]:
+    def get_all(self, model_class) -> list[dict] | None:
         stmt = select(model_class)
 
         with self.session.begin() as session:
             response = session.scalars(stmt).all()
-            data = [obj.to_dict() for obj in response]
+            if response:
+                return [obj.to_dict() for obj in response]
 
-        return data
-
-    def get_one(self, model_class, id: int) -> dict:
-        stmt = select(model_class)
-        stmt.where(model_class.id == id)
+    def get_one(self, model_class, id: int) -> dict | None:
+        stmt = select(model_class).where(model_class.id == id)
 
         with self.session.begin() as session:
             response = session.scalar(stmt)
-            data = response.to_dict()
+            if response:
+                return response.to_dict()
 
-        return data
-
-    def update(self, model_class, id: int, data: dict) -> None:
-        stmt = select(model_class)
-        stmt.where(model_class.id == id)
+    def update(self, model_class, id: int, data: dict) -> bool:
+        stmt = select(model_class).where(model_class.id == id)
         attrs = inspect(model_class).attrs.keys()
 
         with self.session.begin() as session:
             obj = session.scalar(stmt)
-            for key, value in data.items():
-                if key in attrs:
-                    setattr(obj, key, value)
+            if obj:
+                for key, value in data.items():
+                    if key in attrs:
+                        setattr(obj, key, value)
+                return True
+        return False
 
-    def delete(self, model_class, id: int) -> None:
+    def delete(self, model_class, id: int) -> bool:
         with self.session.begin() as session:
-            stmt = select(model_class)
-            stmt.where(model_class.id == id)
+            stmt = select(model_class).where(model_class.id == id)
             obj = session.scalar(stmt)
-            session.delete(obj)
+            if obj:
+                session.delete(obj)
+                return True
+        return False
